backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.13"
dependencies = [
    "ddgs>=9.11.4",
    "fastapi>=0.135.1",
    "google-genai>=1.68.0",
    "passlib[bcrypt]>=1.7.4",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.2",
    "python-jose[cryptography]>=3.5.0",
    "sqlalchemy>=2.0.48",
    "uvicorn>=0.42.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.uv.sources]

[dependency-groups]
dev = [
    "ruff>=0.9.0",
    "mypy>=1.14.0",
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "httpx>=0.27.0",
    "types-python-jose>=3.3.4",
]
